requirements.txt
google-adk
python-dotenv
streamlit
google-generativeai
fpdf